Allow test cases to be run in parallel and allow app servers to
be spawned in parallel to reduce test time from ~99s to ~25s.


From master:
```bash
while go test ./integration -run TestAppServersHA  -count=1 -race; do :; done                                                                                                    
ok  	github.com/gravitational/teleport/integration	98.516s
ok  	github.com/gravitational/teleport/integration	99.719s
ok  	github.com/gravitational/teleport/integration	97.247s
ok  	github.com/gravitational/teleport/integration	97.509s
ok  	github.com/gravitational/teleport/integration	98.961s
ok  	github.com/gravitational/teleport/integration	98.864s
```

With changes in this PR:
```bash
 while go test ./integration -run TestAppServersHA  -count=1 -race; do :; done                                                                                             
ok  	github.com/gravitational/teleport/integration	25.519s
ok  	github.com/gravitational/teleport/integration	25.889s
ok  	github.com/gravitational/teleport/integration	23.359s
ok  	github.com/gravitational/teleport/integration	22.605s
ok  	github.com/gravitational/teleport/integration	23.169s
ok  	github.com/gravitational/teleport/integration	24.411s
```