@rosstimothy 2022-04-26
I further simplified the tests in https://github.com/gravitational/teleport/pull/12128/commits/198d9947f8fc37d1df1ec58fec5cad534e11bef7 to hopefully get CI to pass. This also shaved some more time off the test:

```bash
while go test ./integration -run TestAppServersHA -count=1 -race; do :; done
ok  	github.com/gravitational/teleport/integration	19.709s
ok  	github.com/gravitational/teleport/integration	19.889s
ok  	github.com/gravitational/teleport/integration	20.801s
ok  	github.com/gravitational/teleport/integration	20.360s
ok  	github.com/gravitational/teleport/integration	20.451s
ok  	github.com/gravitational/teleport/integration	19.464s
ok  	github.com/gravitational/teleport/integration	20.956s
ok  	github.com/gravitational/teleport/integration	20.930s
ok  	github.com/gravitational/teleport/integration	19.739s
```